Fix -Wconversion warnings in simulation code. Cast to smaller integer types explicitly. Generally avoid platform-dependent types (size_t) in simulation code. Use float versions of math.h functions, not double.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@10017 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/ShaderProgram.h
@@ -63,7 +63,7 @@ public:
 	{
 		friend class CShaderProgramARB;
 	private:
-		Binding(int v, int f) : vertex(v), fragment(f) { }
+		Binding(int v, int f) : vertex((i16)v), fragment((i16)f) { }
 		i16 vertex;
 		i16 fragment;
 	public:
